+++ server/src/game/behaviours/BugBehaviour.cpp
@@ -20,7 +20,7 @@ namespace rtype::server {
             this->rotation_ = this->getComponent<RotationComponent>();
         }
         this->totalElapsedTime_ += elapsedTime;
-        int step = this->totalElapsedTime_ / this->rate_;
+        long step = this->totalElapsedTime_ / this->rate_;
 
         this->position_->x -= step * this->stepInX_;
         this->position_->y = this->initInY_ + (this->amplitude_ * sin(this->position_->x / this->longitude_));
