+++ server/src/game/behaviours/MissilePlayerBehaviour.cpp
@@ -16,7 +16,7 @@ namespace rtype::server {
             this->position_ = this->getComponent<PositionComponent>();
         }
         this->totalElapsedTime_ += elapsedTime;
-        int step = this->totalElapsedTime_ / this->rate_;
+        long step = this->totalElapsedTime_ / this->rate_;
 
         this->position_->x += step * this->stepInX_;
         if (step)
